Additional updates for lsst.daf.butler subpackage restructuring.
